Fix the gap between thoughts
